Remove copy of cardsavr SDK from the source, refactor and other build related updates.
