+++ package.json
@@ -1,11 +1,12 @@
 {
   "name": "@strivve/strivve-cx",
-  "version": "0.1.1",
+  "version": "0.1.4",
   "private": false,
-  "main": "dist/main.js",
-  "module": "dist/src/index.js",
-  "types": "dist/src/index.d.ts",
+  "main": "main.js",
+  "module": "src/index.js",
+  "types": "src/index.d.ts",
   "dependencies": {
+    "@strivve/strivve-sdk": "^2.12.6",
     "@emotion/react": "^11.10.5",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
@@ -34,7 +35,7 @@
   },
   "scripts": {
     "start": "react-app-rewired start",
-    "build": "rm -rv -f ./dist && tsc && GENERATE_SOURCEMAP=false react-app-rewired build  && cp build/static/js/main*.js dist/main.js && cp build/static/js/main*.js public/main.js",
+    "build": "rm -rv -f ./dist && tsc && GENERATE_SOURCEMAP=false react-app-rewired build  && cp build/static/js/main*.js dist/main.js && cp build/static/js/main*.js public/main.js && cp README.md dist/README.md",
     "test": "react-scripts test",
     "eject": "react-app-rewired eject",
     "format": "prettier --write .",
